Rewrite ificmp{eq,ne} of comparisons during VP
VP (value propagation) will now simplify conditional trees of the
following form into a single fused conditional branch.

    ificmp{eq,ne}
      xcmp{eq,ne}
        lhs
        rhs
      iconst {0,1}

This allows global VP to propagate null/non-null constraints when one
child of the conditional was originally an instanceof, and it was
transformed into a comparison against null earlier in the same pass.

Signed-off-by: Devin Papineau <devinmp@ca.ibm.com>