Add boundary checking for _registerFile
_registerFile is a list of RealRegisters whose indices are
grabbed from RealRegisterEnum.hpp for all RealRegisters. This
enum defines PseudoRealRegisters as well as RealRegisters.
PseudoRealRegisters are not defined to be in the bounds of
_registerFile. Thus, boundary checking needs to be added wherever
needed to prevent one from trying to access aPseudoRealRegister in
_registerFile. Doing so would cause undefined behaviour. This
commit updates the isRealReg and isPseudoReg queries and then
adds boundary checking on _registerFile using these queries.

Signed-off-by: Dhruv Chopra <Dhruv.C.Chopra@ibm.com>