Fixing dcmplt and dcmpgt Instruction handling on PPC
dcmplt and dcmpgt (Along with fcmpgt and fcmplt) use setbool in order to compare and set the required bit results.

Prior, the sequence setbool expands was implemented incorrectly. The result expected was never written.

With this change, I've re-implemented the instruction so that it performs in-line with the expected result. It's notble to explain that this change also contains support to use all conditional register instructions in PPC, which prior to this change did not exist. This was a required prerequisite to using crxor, which was required to reimplement setbool.

Signed-off-by: Alen Badel <Alen.Badel@ibm.com>