Fix _initRanges race between CM and CS
_initRanges is a table that drives parallel setting/reseting of Mark Map
in Tenure/Nursery at the begining of Concurrent Marking CM)
cycle/begining STW phase.

In presence of Concurrent Scavenger (CS), that can over lap with CM, and
if Tenure heap expansion occurs (due to failed tenuring), the table will
be updated. 

This change will make the heap expanding thread simply skip the table
update if it senses that init in is progress. The table update is left
for a later, safer point .

BTW, if initializaiton is already in progress, the thread doing heap
expansion will do Mark Map work for the expaned part of the heap (this
part, and deferral of the table update is actually already handled for
standard Scavenger), so it does not really matter if update of
_initRanges occurs just after Mark Map init
work is complete.

This fix has not direct relation with issue reported by
https://github.com/eclipse/omr/issues/2426. A few conditions are same
(CM and CS overlapping, CS doing Tenure expansion), but the key
difference is what CM is doing: in this scenario Mark Map initialization
vs Card Table cleaning as in the issue.

Some other minor changes:
- split the deferal flag (_rebuildInitWork) into two flags
(expand/contract) for easier potential future debugging (there is a
nasty case of heap contraction that might be compromised)
- renamed a flag to indicates STW phase of Concurernt Mark

Signed-off-by: Aleksandar Micic <amicic@ca.ibm.com>