Cache info if a master handler is registered with a signal
omrsig_protect and omrsig_set_*_async_signal_handler may be invoked
multiple times. If a master handler is already registered with a signal,
then we don't want to re-register the master handler with that signal.

If a master handler is registered with a signal, then this information
is cached in signalsWithMasterHandlers. The port library signal bit is
set in signalsWithMasterHandlers.

When a user overrides the master handler by invoking
omrsig_register_os_handler, then the port library signal bit is un-set
in signalsWithMasterHandlers.

omrsig_set_single_async_signal_handler returns the oldOSHandler. The
oldOSHandler can be the master handler if the user re-invokes
omrsig_set_single_async_signal_handler for a signal. So, a user can
invoke omrsig_register_os_handler using a master handler. In this case,
we must set signalsWithMasterHandlers bits accordingly in
omrsig_register_os_handler.

This allows us to prevent OS calls to re-register master signal
handlers. 

Signed-off-by: Babneet Singh <sbabneet@ca.ibm.com>