Remove unnecessary write barrier
At the end of registerSignalHandlerWithOS, a write barrier is issued
before updating signalsWithHandlers. According to the comment, the write
barrier is needed since registerHandlerMonitor is not acquired before
updating signalsWithHandlers at some places.

It has been confirmed that registerHandlerMonitor is acquired every time
signalsWithHandlers is accessed. So, the write barrier at the end of
registerSignalHandlerWithOS is no longer required. 

In sig_full_shutdown, registerHandlerMonitor is not acquired before
updating signalsWithHandlers. This is an exception since no other code
should execute when OMR is shutting down.

Signed-off-by: Babneet Singh <sbabneet@ca.ibm.com>