Improve aliasToCopyCache performance
aliasToCopyCache is a complicated function which consumes a lot
of time during scavenges. Aliasing provides a significant mutator
throughput improvement so turning it off or changing how much
aliasing happens will degrade performance.

My assertion is that if the current scan cache is already an
aliased copycache then as long as scanCache == copyCache there is
no reason to perform any of the other checks.  This should keep
the same copy order with simpler checks.

Signed-off-by: Charlie Gracie <charlie.gracie@gmail.com>