Remove pass by reference nextScanCache
IncrementalScanCacheBySlot does not need to be passed the
nextScanCache by reference to return the next cache to scan. This
function is marked as void so it can easily return the next cache
to scan as a return value instead of the pass by reference
parameter.

We do not need to store the scanCache in the MM_Environment. This
can easily be passed to all users.

Fix some #if defined tabs

Signed-off-by: Charlie Gracie <charlie.gracie@gmail.com>