Don't define MSVC_RUNTIME_DLL for an invalid MSVC runtime library
omrsig library won't register OS signal handlers on Windows if
MSVC_RUNTIME_DLL points to an invalid MSVC runtime library. If a valid
MSVC runtime library is not found, then the MSVC_RUNTIME_DLL flag should
not be set. If MSVC_RUNTIME_DLL is not set, then compilation will fail.
This will assure that the definition of MSVC_RUNTIME_DLL is updated
whenever OMR is compiled with a newer version of Visual Studio.

Signed-off-by: Babneet Singh <sbabneet@ca.ibm.com>