Handle out-of-order lock acquires and releases
A field named monitor is added in the OMRThreadMCSNode data-structure.
monitor is set to the owner omrthread_monitor_t once the lock is
acquired. When the lock is released, the MCS node with the corresponding
monitor is ejected from the thread's stack.
ASSERT(s) has been replaced with Assert_THR_true(s).

Signed-off-by: Babneet Singh <sbabneet@ca.ibm.com>